requirements.txt
django==3.2.9
flake8==4.0.1
flake8-import-order==0.18.1
pep8-naming==0.12.1
pytest-django==4.4.0
requests==2.26.0